Fix CTR and OFB, plus decrypt group message
We forgot to encrypt the IV in CTR and we forgot to recycle the encrypted IV in OFB. Also in OFB we were using our Support.encrypt() method, which assumes the input is characters. So, I also added an encryptBinary that makes the assumption that the input is already in binary. I use the new method for IV production in both CTR and OFB.